reorganize for unidep package installation
