+++ environment.yml
@@ -1,0 +1,106 @@
+name: frog
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - bcrypt=3.2.0=py312h2bbff1b_1
+  - blas=1.0=mkl
+  - bottleneck=1.3.7=py312he558020_0
+  - brotli=1.0.9=h2bbff1b_8
+  - brotli-bin=1.0.9=h2bbff1b_8
+  - bzip2=1.0.8=h2bbff1b_6
+  - ca-certificates=2024.7.2=haa95532_0
+  - cffi=1.16.0=py312h2bbff1b_1
+  - contourpy=1.2.0=py312h59b6b97_0
+  - cryptography=42.0.5=py312h89fc84f_1
+  - cycler=0.11.0=pyhd3eb1b0_0
+  - expat=2.6.2=hd77b12b_0
+  - fonttools=4.51.0=py312h2bbff1b_0
+  - freetype=2.12.1=ha860e81_0
+  - icc_rt=2022.1.0=h6049295_2
+  - icu=73.1=h6c2663c_0
+  - intel-openmp=2023.1.0=h59b6b97_46320
+  - jpeg=9e=h827c3e9_3
+  - kiwisolver=1.4.4=py312hd77b12b_0
+  - krb5=1.20.1=h5b6d351_0
+  - lcms2=2.12=h83e58a3_0
+  - lerc=3.0=hd77b12b_0
+  - libbrotlicommon=1.0.9=h2bbff1b_8
+  - libbrotlidec=1.0.9=h2bbff1b_8
+  - libbrotlienc=1.0.9=h2bbff1b_8
+  - libclang=14.0.6=default_hb5a9fac_1
+  - libclang13=14.0.6=default_h8e68704_1
+  - libdeflate=1.17=h2bbff1b_1
+  - libexpat=2.6.2=h63175ca_0
+  - libffi=3.4.4=hd77b12b_1
+  - libpng=1.6.39=h8cc25b3_0
+  - libpq=12.17=h906ac69_0
+  - libsqlite=3.46.0=h2466b09_0
+  - libtiff=4.5.1=hd77b12b_0
+  - libusb=1.0.27=hcfcfb64_100
+  - libwebp-base=1.3.2=h2bbff1b_0
+  - libzlib=1.2.13=h2466b09_6
+  - llvmlite=0.43.0=py312hf2fb9eb_0
+  - lz4-c=1.9.4=h2bbff1b_1
+  - matplotlib=3.8.4=py312haa95532_0
+  - matplotlib-base=3.8.4=py312hc7c4135_0
+  - mkl=2023.1.0=h6b88ed4_46358
+  - mkl-service=2.4.0=py312h2bbff1b_1
+  - mkl_fft=1.3.8=py312h2bbff1b_0
+  - mkl_random=1.2.4=py312h59b6b97_0
+  - numba=0.60.0=py312h0158946_0
+  - numexpr=2.8.7=py312h96b7d27_0
+  - numpy=1.26.4=py312hfd52020_0
+  - numpy-base=1.26.4=py312h4dde369_0
+  - openjpeg=2.5.2=hae555c5_0
+  - openssl=3.3.1=h2466b09_2
+  - packaging=24.1=py312haa95532_0
+  - pandas=2.2.2=py312h0158946_0
+  - paramiko=2.8.1=pyhd3eb1b0_0
+  - pillow=10.4.0=py312h827c3e9_0
+  - pip=24.2=py312haa95532_0
+  - plumbum=1.8.3=pyhd8ed1ab_0
+  - ply=3.11=py312haa95532_1
+  - pybind11-abi=5=hd3eb1b0_0
+  - pycparser=2.21=pyhd3eb1b0_0
+  - pylablib=1.4.3=py312h4389bb4_0
+  - pynacl=1.5.0=py312h8cc25b3_0
+  - pyparsing=3.0.9=py312haa95532_0
+  - pyqt=5.15.10=py312hd77b12b_0
+  - pyqt5-sip=12.13.0=py312h2bbff1b_0
+  - pyqtgraph=0.13.1=py312haa95532_0
+  - pyserial=3.5=py312haa95532_1
+  - python=3.12.3=h2628c8c_0_cpython
+  - python-dateutil=2.9.0post0=py312haa95532_2
+  - python-tzdata=2023.3=pyhd3eb1b0_0
+  - python_abi=3.12=5_cp312
+  - pytz=2024.1=py312haa95532_0
+  - pyusb=1.2.1=pyhd8ed1ab_0
+  - pyvisa=1.14.1=pyhd8ed1ab_0
+  - pywin32=305=py312h2bbff1b_0
+  - pywin32-on-windows=0.1.0=pyh07e9846_2
+  - qt-main=5.15.2=h19c9488_10
+  - rpyc=5.3.0=pyhd8ed1ab_0
+  - scipy=1.13.1=py312hbb039d4_0
+  - seabreeze=2.9.2=py312h275cf98_0
+  - setuptools=72.1.0=py312haa95532_0
+  - sip=6.7.12=py312hd77b12b_0
+  - six=1.16.0=pyhd3eb1b0_1
+  - sqlite=3.45.3=h2bbff1b_0
+  - tbb=2021.8.0=h59b6b97_0
+  - tk=8.6.14=h0416ee5_0
+  - tornado=6.4.1=py312h827c3e9_0
+  - typing_extensions=4.11.0=py312haa95532_0
+  - tzdata=2024a=h04d1e81_0
+  - ucrt=10.0.20348.0=haa95532_0
+  - unicodedata2=15.1.0=py312h2bbff1b_0
+  - vc=14.40=h2eaa2aa_0
+  - vc14_runtime=14.40.33810=ha82c5b3_20
+  - vs2015_runtime=14.40.33810=h3bf8584_20
+  - wheel=0.43.0=py312haa95532_0
+  - xz=5.4.6=h8cc25b3_1
+  - zlib=1.2.13=h2466b09_6
+  - zstd=1.5.5=hd43e919_2
+  - pip:
+      - pyft232==0.12
+prefix: C:\Users\Splinter\.conda\envs\frog
