+++ src/frogware_fcxqm.egg-info/SOURCES.txt
@@ -1,0 +1,27 @@
+README.md
+pyproject.toml
+src/frogware_fcxqm/BBO.py
+src/frogware_fcxqm/Error.py
+src/frogware_fcxqm/Gui_Controller.py
+src/frogware_fcxqm/PhaseRetrieval.py
+src/frogware_fcxqm/PlotAndTableFunctions.py
+src/frogware_fcxqm/PullDataFromOSA.py
+src/frogware_fcxqm/QRC_file_rc.py
+src/frogware_fcxqm/Window.py
+src/frogware_fcxqm/__init__.py
+src/frogware_fcxqm/python_phase_retrieval.py
+src/frogware_fcxqm/scripts.py
+src/frogware_fcxqm/simulation_header.py
+src/frogware_fcxqm.egg-info/PKG-INFO
+src/frogware_fcxqm.egg-info/SOURCES.txt
+src/frogware_fcxqm.egg-info/dependency_links.txt
+src/frogware_fcxqm.egg-info/entry_points.txt
+src/frogware_fcxqm.egg-info/top_level.txt
+src/frogware_fcxqm/hardware_comms/APT.py
+src/frogware_fcxqm/hardware_comms/Emulators.py
+src/frogware_fcxqm/hardware_comms/MotorClassFromAptProtocolConnor.py
+src/frogware_fcxqm/hardware_comms/OceanOpticsSpectrometer.py
+src/frogware_fcxqm/hardware_comms/ThorlabsKinesisMotor.py
+src/frogware_fcxqm/hardware_comms/device_interfaces.py
+src/frogware_fcxqm/hardware_comms/stellarnet_peter.py
+src/frogware_fcxqm/hardware_comms/utilities.py
