+++ src/frogware_fcxqm.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[gui_scripts]
+frogware = frogware_fcxqm:scripts.frogware
