+++ src/frogware_fcxqm/Error.py
