+++ src/frogware_fcxqm/hardware_comms/ThorlabsKinesisMotor.py
@@ -1,6 +1,6 @@
 from hardware_comms.device_interfaces import LinearMotor, StageOutOfBoundsException
 from pylablib.devices.Thorlabs import KinesisMotor
-from utilities import T_fs_to_dist_um, dist_um_to_T_fs
+from .utilities import T_fs_to_dist_um, dist_um_to_T_fs
 
 '''
 Generic class for all Thorlabs linear motors which
